rh-utils: add script find-backports.sh
When backporting commits, we should do `git cherry-pick -x`,
so that the commit message contains:
    (cherry picked from commit $SHA)

Similarly, when fixing an actual bug of a previous commit, we
mark the faulty commit with:
    Fixes: $SHA

This script walks through a commit range of a stable branch,
and finds all $BACKPORTED_COMMITS (the sources from where we
cherry-picked).
Then it walks through a commit range of upstream/unstable branch,
and searches for $FIXES of the $BACKPORTED_COMMITS.

Then it prints all $FIXES, that are not yet backported. You probably
should backport those commits as well.