platform: get rid of old logging macros
The naming of these logging macros is unexpected, as we use such
macros only here in platform.

For these messages we cannot use the default _LOGD() set of macros,
because there is no @platform instance around. So let's introduce an
alternative set of logging macros (_LOG2D(), etc) and use it.