contrib/rh-bkr: add dummy RPM provider "none"
If you don't want to install any RPMs, this will lead to an
error sustituing $RPM_LIST template.

Add a dummy provider that can be used:

  ./bkr.py submit -r none ...

You are still expected to explicitly pass the -r argument.
We don't want to interpret a missing argument as "none" by
default, because it's a rather uncommon usecase.