wifi/ap: use direct-hashing for aps hash
The @aps hash has the D-Bus path of the exported
object as key. It already rightly saved to additionally
copy the string and relied on the path being stable.

When doing that, we can just go one step further and
use direct-hashing instead of string-hashing.

Note that NMExportedObject already promises that
the path will not change as long as the object is
exported. See code comments in the export/unexport
functions.