wifi/ap: set the current-ap after adding the new AP
First add the new AP, before setting it as current.

Also set the AP *after* thawing the notifications. Otherwise
it is not clear which notification gets raised first as their
order is undefined. But we want that the client first sees
the new AP and later gets a notification about having a new
current.