contrib/rh-bkr: refactor RPM scheme handlers in bkr.py
Get rid of UploadFileSsh. This was hardly used because it
is cumbersome to use. Also, the default scheme was file://
if no scheme was given, which makes this especially non-useful.

Replace it by a new provider RPM, so that you can just specify
any rpms you want to install, like --rpm package1,package2.

Also, drop the no longer needed init() and prepare() steps.

Finally, rename UploadFile* classes to RpmScheme*.